Seems that test db values are not saving
